moved functions related to update_cal_data.py to new file
