+++ update_cal_data.py
@@ -1,8 +1,7 @@
 ##!/home/ubuntu/miniconda2/envs/cal_service/bin/python
 # encoding: utf-8
-import app
+import functions as app
 import simplejson as json
-
 
 # Gets lists of meetings, concatenates meeting lists from all sources, then saves json locally.
 def refresh_all_meetings():
@@ -10,7 +9,6 @@ def refresh_all_meetings():
     meetup_events = app.get_meetup_events(group_lists['Meetup.com'])
     eventbrite_events = app.get_eventbrite_events(group_lists['Eventbrite.com'])
     eventbrite_venues = app.get_eventbrite_venues(eventbrite_events)
-    
     events = (
         app.format_meetup_events(meetup_events, group_lists['Meetup.com'])
         + app.format_eventbrite_events(events_list=eventbrite_events,
